backend/requirements.txt
# dependencies for backend
fastapi
uvicorn
langchain
langchain-community
langchain-chroma
langchain-google-genai
chromadb
langchain_groq
python-multipart
pypdf
sentence-transformers
python-dotenv